sentiment_analysis/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
psycopg2-binary==2.9.2
requests==2.26.0
urllib3==1.26.7
vaderSentiment==3.3.2
